+++ src/main/kotlin/com/widehouse/dnd/character/action/AttackResult.kt
@@ -1,4 +1,6 @@
-package com.widehouse.dnd.character
+package com.widehouse.dnd.character.action
+
+import com.widehouse.dnd.character.Creature
 
 class AttackResult(val target: Creature, private val damage: Int) {
     fun resolve() {
